back-end/requirements.txt
# ----------------------------------------
# CORE FRAMEWORK DEPENDENCIES (Install first - most stable)
# ----------------------------------------
Flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ----------------------------------------
# SYSTEM UTILITIES (Small, stable packages)
# ----------------------------------------
click>=8.1.3
python-dotenv==1.0.0
requests>=2.32.3
python-multipart==0.0.6
PyJWT==2.8.0

# ----------------------------------------
# CORE AI/ML LIBRARIES (Large but essential)
# ----------------------------------------
pydantic>=2.8.2

# ----------------------------------------
# SPECIALIZED LIBRARIES (Order by size/compile time)
# ----------------------------------------

# PDF Processing (lightweight first)
PyPDF2>=3.0.1
pdfplumber>=0.7.6
PyMuPDF==1.23.8
reportlab>=4.0.0

# Audio (requires compilation)
SpeechRecognition==3.10.0
PyAudio==0.2.11

# Google Services - Let pip resolve compatible versions
google-generativeai>=0.8.0
firebase-admin>=6.2.0

# Database
pymongo>=4.10.1
aiofiles>=24.1.0

# Cloud Storage
boto3>=1.36.0

# ----------------------------------------
# HEAVY ML/AI LIBRARIES (Install last)
# ----------------------------------------
langchain>=0.3.14
langchain-core>=0.3.30
langchain-mongodb>=0.4.0
langchain-openai>=0.3.1
langchain-huggingface>=0.1.2

# AI Agents
smolagents>=1.4.1

# Monitoring
opik>=0.1.0
comet_ml>=3.47.6

# Development (Optional - consider removing for production)
ipykernel>=6.29.5
loguru>=0.7.3
pydantic-settings>=2.7.0